GEODE-8783: Small changes after review

+++ geode-wan/src/distributedTest/java/org/apache/geode/internal/cache/wan/WANTestBase.java
@@ -1252,7 +1252,7 @@ public class WANTestBase extends DistributedTestCase {
     stats.add(statistics.getSecondaryEventQueueSize());
     stats.add(statistics.getEventsProcessedByPQRM());
     stats.add(statistics.getEventsExceedingAlertThreshold());
-    stats.add(statistics.getBatchesWithIncompleteTransactions());
+    stats.add((int) statistics.getBatchesWithIncompleteTransactions());
     return stats;
   }
